--- a/.devops/templates/build-test-lint.yml
+++ b/.devops/templates/build-test-lint.yml
@@ -18,8 +18,8 @@ steps:
       # @fluentui/api-docs is used within apps/public-docsite-resources/just.config.ts,
       # @fluentui/digest is used within packages/fluentui/perf-test-northstar/just.config.ts, thus it needs to be build in advance
       #
-      # NOTE: we are running this via lage in order to get cache hits later on
-      yarn lage build --to @fluentui/api-docs @fluentui/digest --verbose
+      # NOTE: we are running this via nx in order to get cache hits later on
+      yarn nx run-many -t build -p api-docs digest --verbose
       yarn tsc -p ./tsconfig.just-scripts-configs.json
     displayName: Type-check just.config.ts files