--- a/.devops/templates/deployE2E.yml
+++ b/.devops/templates/deployE2E.yml
@@ -16,11 +16,11 @@ steps:
 
   # this also builds FUI N* docs if appropriate
   - script: |
-      yarn nx affected -t bundle --nxBail --verbose $(sinceArg)
+      yarn nx affected -t bundle --nxBail $(sinceArg)
     displayName: bundle
 
   - script: |
-      yarn nx affected -t build-storybook --nxBail --verbose $(sinceArg)
+      yarn nx affected -t build-storybook --nxBail $(sinceArg)
     displayName: build Storybooks
 
   ## This runs regardless of scope, the app will adapt to the scope as well
@@ -54,7 +54,7 @@ steps:
       githubTargetLink: $(deployUrl)/
 
   - script: |
-      yarn nx affected -t e2e --nxBail --verbose $(sinceArg)
+      yarn nx affected -t e2e --nxBail $(sinceArg)
     displayName: Cypress E2E tests
 
   - template: cleanup.yml