--- a/.devops/templates/deployE2E.yml
+++ b/.devops/templates/deployE2E.yml
@@ -53,10 +53,9 @@ steps:
       # This MUST have a trailing slash, or the links to PR deploy site assets won't work
       githubTargetLink: $(deployUrl)/
 
+  # TODO: once https://github.com/nrwl/nx/issues/26903 will be fixed we can remove NX_PARALLEL=1 override via env variable
   - script: |
       NX_PARALLEL=1 yarn nx affected -t e2e --nxBail --parallel 1 $(sinceArg)
     displayName: Cypress E2E tests
-    env:
-      NX_PARALLEL: 1
 
   - template: cleanup.yml