--- a/.devops/templates/deployE2E.yml
+++ b/.devops/templates/deployE2E.yml
@@ -56,5 +56,7 @@ steps:
   - script: |
       yarn nx affected -t e2e --nxBail --parallel 1 $(sinceArg)
     displayName: Cypress E2E tests
+    env:
+      NX_PARALLEL: 1
 
   - template: cleanup.yml