--- a/.devops/templates/variables.yml
+++ b/.devops/templates/variables.yml
@@ -16,7 +16,9 @@ variables:
   # Prevents failures on CI when "caniuse-lite" becomes outdated
   BROWSERSLIST_IGNORE_OLD_DATA: true
 
-  NX_PARALLEL: 3
+  NX_PARALLEL: 8
+  NX_PREFER_TS_NODE: true
+  NX_VERBOSE_LOGGING: true
 
   # Also accessed as process.env.DEPLOYHOST
   deployHost: 'fluentuipr.z22.web.core.windows.net'