--- a/.storybook/main.js
+++ b/.storybook/main.js
@@ -80,7 +80,7 @@ module.exports = /** @type {Omit<StorybookConfig,'typescript'|'babel'>} */ ({
   webpackFinal: config => {
     registerTsPaths({ config, configFile: tsConfigPath });
 
-    if ((process.env.CI || process.env.TF_BUILD || process.env.LAGE_PACKAGE_NAME) && config.plugins) {
+    if ((process.env.CI || process.env.TF_BUILD) && config.plugins) {
       // Disable ProgressPlugin in PR/CI builds to reduce log verbosity (warnings and errors are still logged)
       config.plugins = config.plugins.filter(value => value && value.constructor.name !== 'ProgressPlugin');
     }