--- a/apps/perf-test-react-components/project.json
+++ b/apps/perf-test-react-components/project.json
@@ -5,6 +5,7 @@
   "implicitDependencies": [],
   "tags": ["vNext"],
   "targets": {
+    "test-perf": { "dependsOn": ["bundle"] },
     "bundle": { "dependsOn": [] }
   }
 }