--- a/apps/public-docsite-resources/project.json
+++ b/apps/public-docsite-resources/project.json
@@ -1,7 +1,7 @@
 {
   "name": "public-docsite-resources",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "application",
+  "projectType": "library",
   "implicitDependencies": [],
   "tags": ["v8"]
 }