--- a/apps/ssr-tests/project.json
+++ b/apps/ssr-tests/project.json
@@ -5,6 +5,9 @@
   "implicitDependencies": [],
   "tags": ["v8"],
   "targets": {
-    "test-ssr": { "dependsOn": ["bundle", "^bundle"] }
+    "test-ssr": {
+      "dependsOn": ["bundle"]
+    },
+    "bundle": { "dependsOn": ["^build"] }
   }
 }