--- a/apps/vr-tests-web-components/.storybook/main.cjs
+++ b/apps/vr-tests-web-components/.storybook/main.cjs
@@ -75,7 +75,7 @@ module.exports = /** @type {import('../../../.storybook/types').StorybookBaseCon
     );
 
     // Disable ProgressPlugin which logs verbose webpack build progress. Warnings and Errors are still logged.
-    if (process.env.TF_BUILD || process.env.LAGE_PACKAGE_NAME) {
+    if (process.env.TF_BUILD) {
       config.plugins = config.plugins.filter(({ constructor }) => constructor.name !== 'ProgressPlugin');
     }