--- a/azure-pipelines-pr.yml
+++ b/azure-pipelines-pr.yml
@@ -8,7 +8,7 @@ trigger: none
 variables:
   - ${{ if not(startsWith(variables['Build.SourceBranch'], 'refs/heads/')) }}:
       - name: sinceArg
-        value: --since $(targetBranch)
+        value: --base $(targetBranch)
 
   - group: fabric-variables