--- a/azure-pipelines.bundlesize.yml
+++ b/azure-pipelines.bundlesize.yml
@@ -25,7 +25,7 @@ jobs:
         displayName: yarn
 
       - script: |
-          yarn nx affected -t bundle-size --nxBail --verbose $(sinceArg)
+          yarn nx affected -t bundle-size --nxBail $(sinceArg)
         displayName: build packages & create reports
 
       - script: |