--- a/scripts/fluentui-publish/jest.config.js
+++ b/scripts/fluentui-publish/jest.config.js
@@ -10,6 +10,7 @@ module.exports = {
     '^.+\\.tsx?$': ['@swc/jest', {}],
   },
   coverageDirectory: './coverage',
+  testTimeout: 20000,
   testEnvironment: 'node',
   setupFiles: ['<rootDir>/jest-setup.js'],
 };