--- a/tools/workspace-plugin/src/generators/react-component/index.ts
+++ b/tools/workspace-plugin/src/generators/react-component/index.ts
@@ -22,9 +22,9 @@ export default async function (tree: Tree, schema: ReactComponentGeneratorSchema
 
   return () => {
     const root = workspaceRoot;
-    const { npmPackageName, componentName } = options;
+    const { npmPackageName, project, componentName } = options;
 
-    execSync(`yarn nx run ${npmPackageName}:generate-api`, {
+    execSync(`yarn nx run ${project}:generate-api`, {
       cwd: root,
       stdio: 'inherit',
     });