--- a/typings/environment/index.d.ts
+++ b/typings/environment/index.d.ts
@@ -2,7 +2,6 @@ declare namespace NodeJS {
   interface ExtendedProcessEnv {
     // env: {
     NODE_ENV?: 'production' | 'development' | 'test';
-    LAGE_PACKAGE_NAME?: string;
     CI?: string;
     TF_BUILD?: string;
   }