<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

- `lage` is used for task orchestration
- custom `runPublished` script is used to run targets within release pipelines

## New Behavior

- `nx` is used for any kind of task orchestration including releases
- `runPublished` has been removed
- `lage` package  has been removed from dependencies and any related logic has been refactored to use `nx` under the hood
- added missing tags to `project.json` in order to be able to filter projects when running releases

### DX changes

**`yarn buildto`**

```diff
-yarn lage <target> --to <project-name>
+yarn nx run <project-name>:<target>
```

**running task on affected projects by your changes**

```diff
-yarn lage <target> --since=master
+yarn nx affected -t <target> --base=master
```

**running tasks on whole repo**

```diff
-yarn lage <target>
+yarn nx run-many -t <target>
```

_💡 Note:_ `run-many` provides powerful filtering capabilities https://nx.dev/nx-api/nx/documents/run-many#examples

**CI logs**

_Before:_

- mismatch of task execution

![image](https://github.com/user-attachments/assets/f0c6ebf6-5b89-4ad5-bca3-e0bb62927660)

- hard to troubleshoot failures

<img width="1248" alt="image" src="https://github.com/user-attachments/assets/b8ebfe07-8c01-4a3e-ab23-c3b72b41f9b0">



_After:_

- task execution output is aggregated per task ( while behind the scenes things run in parallel )

<img width="1090" alt="image" src="https://github.com/user-attachments/assets/b58596c4-90d4-4621-8c75-0d80b8fba51e">

- easy to troubleshoot

![image](https://github.com/user-attachments/assets/1bf10b20-cf78-4674-a864-d44252896300)


### Pipeline perf

#### `build-test-lint`

| Before | After |
|--------|--------|
| 46m 20s | 46m 7s / **𝚫 0.5% FASTER** |

#### `e2e-deploy`

| Before | After |
|--------|--------|
| 32m 31s | 26m 50s / **𝚫 17% FASTER** |

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Implements partially https://github.com/microsoft/fluentui/issues/30267
- closes https://github.com/microsoft/fluentui/issues/26147,
- closes https://github.com/microsoft/fluentui/issues/31920,
- closes https://github.com/microsoft/fluentui/pull/31710, 
- closes https://github.com/microsoft/fluentui/issues/21686

### NX issues

- [x] https://github.com/nrwl/nx/issues/26903 / has workaround
- [ ] https://github.com/nrwl/nx/discussions/26899 / needed for simplified commands definition on CI
- [ ] https://github.com/nrwl/nx/discussions/26901  / needed for faster CI execution as less projects will be invoked
- [x] https://github.com/nrwl/nx/issues/27143 / DX improvement