@fabricteam 2024-07-09
## 📊 Bundle size report

✅ No changes found
<div id="bundleSizeComment9423"/>